Thu, Jan 27, 2022, 5:35 PM +00:00

+++ readability/readability.c
@@ -14,8 +14,8 @@ int main(void)
     int words = count_words(text);
     int sentences = count_sentences(text);
 
-    float L = 100 * letters / words;
-    float S = 100 * sentences / words;
+    double L = 100 * letters / words;
+    double S = 100 * sentences / words;
 
     int index = round(0.0588 * L - 0.296 * S - 15.8);
     if (index < 1)
